requirements.txt
requests
beautifulsoup4
discord-py-interactions
discord
python-decouple